Add images, slideshow styles for simple posts
The slideshow mode allows to use the post as a support for short
presentations, by making the text and adding extra space between
the images.

That mode can be enabled on demand by adding the `slideshow` class
to the post <body> tag (e.g. from the web browser inspector console).